feat: Update Compliance Manager MQTT Mock
The compliance manager mock in the MQTT library has been updated. Changes include the removal of some lodash import and various types from coldpbc interfaces. This was done to enhance efficiency and promote cleaner code. Modifications were also made to the responses of some functions, with additional payload returned and the format of existing payloads changed.