feat: Refactor search parameter removal in Questionnaire components
The code for removing query parameters in the QuestionnaireContainer and QuestionnaireQuestionItem components has been updated. Instead of resetting the entire list of parameters, it now specifically removes the 'section' and 'question' parameters from the list. This approach provides a better control over query parameters and prevents loss of other unrelated search parameters.