feat: Update handling of undefined data and modify mocks
The code now properly handles cases where the ordered data in the 'complianceManagerOverviewSectionGroup' component is undefined, showing a spinner in this case. The mocks related to 'mqtt' and 'complianceManager' were also modified for test data consistency. Additionally, the conditional checks in 'complianceManagerOverviewSection' have been updated to account for ordered question changes.