feat: Update logging in coldMQTTProvider
The logging in the coldMQTTProvider has been updated to provide more accurate information. Previously, the key value was used in log messages when receiving and ignoring messages from IOT, but it has been changed to show the topic instead. This change ensures the logs reflect the accurate source of the message.