feat: Refactor compliance data fetching and image URL setting
Simplified the handling of compliance data by directly using data from complianceSWR. A default image URL has also been added for cases when the `image_url` property does not exist in the compliance data. This mitigates the risk of a broken image link and optimizes the way we fetch and handle compliance data.