feat: Add scoring functionality to questionnaire components
The commit introduces an answer scoring feature to the React components used in questionnaires. This includes 'QuestionnaireSelect', 'QuestionnaireYesNo', and 'QuestionnaireQuestionItem'. The changes also include adding the lodash 'get' method where necessary. Any selected answers can now be scored according to a mapping of answers and scores, elevating the functionality of the questionnaire feature.