feat: Refactor compliance manager components and improve error handling
The commit includes changes to several components in the compliance manager section. It refactors overview widgets to better handle data using lodash 'get' and remove unnecessary imports. It also introduces improvements to the error state UI and error handling in the storybook of the compliance manager. Furthermore, it also amended visual changes in the status card component.