feat: Add Suppliers interface and update SuppliersDataGrid
A new interface for Suppliers was created in the 'interfaces' folder. In addition, the SuppliersDataGrid component was updated to fetch and display supplier data dynamically from an API instead of using hardcoded mock data. Applied theme updates in 'main.tsx' and 'preview.mjs'. Mocks for certifications and suppliers were also updated.